requirements.txt
Django==4.2
django-filter==21.1
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
Pillow==9.4.0
psycopg2-binary==2.8.4
python-dateutil==2.8.2
python-dotenv==1.0.0
django-jazzmin==2.6.1
mixer==7.2.2
coverage==7.6.1